Remove -Wuseless-cast for GCC builds
It gave a warning in 32-bit mode in the following code, in
include_guard/macro_name.cpp:

  auto const          m = static_cast< unsigned >( alnum.length() - 1 )

The reason is that std::size_t, the type of alnum.length() - 1, and
unsigned were synonyms. But the cast was necessary in 64-bit mode. Since
this seemed difficult to silence without obfuscating the code, I gave up
using the warning option.