Don't explicitly compare the result of <cctype>'s "predicate" functions
This reverts a commit done on August 4, 2018. In fact, although we
adhere to the "explicit is better than implicit" and "say what you mean,
mean what you say" principles, cfr.

  if ( std::isprint( c ) )

to

  if ( std::isprint( c ) != 0 )

Here, "is print(able)?" reads better then "does isprint() return
non-null?"

The negative case is equally, if not more, interesting: cfr.

  if ( ! std::isprint( c ) )

to

  if ( std::isprint( c ) == 0 )

Note that there's a usage of std::isdigit() in luhn.cpp consistent with
our new choice (it was inconsistent before).

Also, as James noticed, the reason these functions return an int is that
they predated the introduction of the bool type. Conceptually, though,
they do return a boolean.