Capitalize a Microsoft header name (<wincrypt.h>) as in the filesystem
We decided to capitalize the Microsoft header names as they are in the
documentation (commit done on November 23, 2016). But Clang warns when
they differ in case from the corresponding filesystem file. So we'll
capitalize them as required by Clang (whether that coincides with the
documentation or not ---it does, in this case AFAICS).

Note, too, that the new rule is better in that we could not be sure that
the same header name is capitalized the same way in different pages of
the documentation (I hope the files are consistent across different
installations, instead).