Add a static_assert for endian_codec
I was hesitant to add this, because, probably, the cryptography parts
still compile as C++03, and someone might want to extract them to use
the code with older compilers. But we are no longer compiling as C++03,
so we can't guarantee that possibility anyway. (And removing/replacing a
static_assert is easy.)