Reimplement uuid by using a single array data member
Reason: this simplifies the equality and relational operators, which I
want to introduce. It also simplifies a possible copy_to() member
function. It just slightly complicates the constructor from variant and
version. (The stream inserter has more or less the same complexity but
gets rid of the problem of choosing a) a width b) a strategy to output
unsigned chars as numbers (unary +, binary +, or static_cast).)

Note that when constructing 'clock_seq' it's enough to generate *three*
random numbers; I noticed this when reimplementing, and just removed the
generation of the fourth random number.