Add a missing 'virtual' keyword
Note: at least for the moment, we don't follow the Core Guidelines here;
I see no danger, in fact, in having 'virtual' together with 'override',
and 'override' appears at the end of the declaration, possibly after
other keywords ('const' and 'noexcept'), so might be overlooked.

Note 2: I know that the program option facilities are inconsistent in
this respect.