Remove a facility: accumulate()
Reason: unfortunately, I didn't note the rationale for designing
breath::accumulate(), but it was probably optimization with respect to
std::accumulate() for merkle_damgard_machine's (hashers): see the
comments to merkle_damgard_machine introduced with this commit.

But, at least for the case of hashers, you could rely on the hasher copy
assignment operator to do nothing on self-assignment, or just use
std::for_each().

So, unless I discover that the rationale for breath::accumulate() was
*not* efficiency, I'm going to remove it.

PS: another important point is that breath::accumulate()'s design means
that the operation used is coupled to the type. If, for instance, you
wanted to accumulate over a range of BigInt values in two different
ways, one using addition and one using multiplication, you'd need to
write another algorithm (functors are more versatile than template
specialization in this context). And this is another reason to remove
breath::accumulate().