Disable Markdown support in the Doxygen configuration
Reason: when it's enabled, with Doxygen 1.8.15, the whole paragraph "A
note about std::accumulate()", in the documentation of
merkle_damgard_machine, is being rendered as a gigantic code block.
Note that disabling it doesn't solve all the issues either (the code
block of the function-call operator is rendered on one line).