Re-add -Wuseless-cast for GCC builds
The option was removed because it gave a warning in macro_name.cpp which
I found difficult to avoid (see the commit "Remove -Wuseless-cast for
GCC builds"). But I found an easy way, now, to do that: just cast to
int. This works because of the particular choice of type for
entropy_source::result_type: if e.g. that was unsigned short, we'd get a
-Wconversion warning when passing the int to entropy_source's
operator()(). Thus, I might need to tweak the code again.