Make the output directory a target of its own
Reason: besides being more logical and using Make properly (we were
using it as an imperative language, before), this gets rid of the
problem that if e.g. the user misspelled the compiler name on the
command line (say, he wrote 'compiler=clong system=windows make') we
would create a directory with the wrong name.