Include start.mk after defining $(source_files)
Reason: if $(source_files) is not defined in start.mk, then the include
of the .dep files fails; which means that all the automatic dependency
management is broken. Yes, this is a bug fix.

I noticed the bug thanks to --warn-undefined-variables, recently
introduced, but I should have checked that the dependency on .hpp's
worked (it didn't). :-(